rearranged main function
