+++ firmware/main/CMakeLists.txt
@@ -1,4 +1,5 @@
 idf_component_register(SRCS
+    "main.c"
     "tcp_server.c"
     "command.c"
     "motor.c"
