+++ firmware/main/command.h
@@ -1,0 +1,9 @@
+#ifndef COMMAND_H
+#define COMMAND_H
+
+#include "freertos/FreeRTOS.h"
+#include "freertos/task.h"
+void command_task(void *pvParameters);
+
+#endif
+
